Merged PR 25468: Pipeline rule to check failure count
This PR includes:
1. Variable group added to apigee and trader pipelines
2. Check to compare the failed count with previous failed count using variable values

For storing the variable values, we are not using static variable. This is a dynamic variable for which value will be updated for every run. Variables are maintained here.
https://wowonline.visualstudio.com/Mobile/_apps/hub/ms.vss-distributed-task.hub-library?itemType=VariableGroups&view=VariableGroupView&variableGroupId=717&path=MobileAuto-UAT

Apigee_UAT_LastFailCount: variable will be updated if the run is for apigee in UAT env
Trader_UAT_LastFailCount: variable will be updated if the run is for trader in UAT env