Merged PR 30317: Implementation for Mutation - CreateList
### Implementation for Mutation for CreateList - More details in Jira

There are 3 ways to do it,

The simplest and most matching way to query is to make each variable separately defined and then pass it as a string and into to mutation (less code change). - we will continue with this one.

the second way is to make a model(object) of variable type and then pass it - we are not going to continue due to the overhead of creating an extra model type

### Link to JIRA ticket with brief summary: [VEGDX-1569](https://woolworthsdigital.atlassian.net/browse/VEGDX-1569)

### Screenshots:![image.png](https://wowonline.visualstudio.com/2c47de12-5d5b-435d-8005-74afeb496636/_apis/git/repositories/7cb3b87c-785a-4d60-b22a-bd54c1480cf6/pullRequests/30317/attachments/image.png)

### Does this PR contain a reusable code (or potentially can be moved to a standalone function)? Use this as an example for Mutation